+++ lib/views/auth_views/update_profile_page.dart
@@ -2,13 +2,19 @@
 ///
 /// a page where user can update their profile settings
 // TODO: edit this page
+// ignore_for_file: use_build_context_synchronously
+
 import 'package:flutter/material.dart';
-import 'package:firebase_auth/firebase_auth.dart';
+import 'package:form_builder_validators/form_builder_validators.dart';
 import 'package:image_picker/image_picker.dart';
+import 'package:loader_overlay/loader_overlay.dart';
 import 'package:universal_io/io.dart';
 import 'package:xiaokeai/components/main_view.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:xiaokeai/components/quick_settings_menu.dart';
+import 'package:xiaokeai/helpers/dependencies_injection.dart';
+import 'package:xiaokeai/services/auth/acc/auth_wrapper.dart';
+import 'package:xiaokeai/shared/views/ui_consts.dart';
 
 class UpdateProfilePage extends StatefulWidget {
   const UpdateProfilePage({super.key});
@@ -18,6 +24,7 @@ class UpdateProfilePage extends StatefulWidget {
 }
 
 class _UpdateProfilePageState extends State<UpdateProfilePage> {
+  final AuthWrapper _authWrapper = getIt<AuthWrapper>();
   final _formKey = GlobalKey<FormState>();
   final _displayNameController = TextEditingController();
   File? _image;
@@ -25,8 +32,8 @@ class _UpdateProfilePageState extends State<UpdateProfilePage> {
   @override
   void initState() {
     super.initState();
-    _displayNameController.text =
-        FirebaseAuth.instance.currentUser?.displayName ?? '';
+    _authWrapper.refreshAttributes();
+    _displayNameController.text = _authWrapper.displayName;
   }
 
   Future<void> _getImage() async {
@@ -50,71 +57,96 @@ class _UpdateProfilePageState extends State<UpdateProfilePage> {
       body: Container(
         decoration: BoxDecoration(
           gradient: LinearGradient(
-            begin: Alignment.topRight,
-            end: Alignment.bottomLeft,
-            colors: [Colors.orange[400]!, Colors.red[500]!],
+            begin: Alignment.topCenter,
+            end: Alignment.bottomCenter,
+            colors: [
+              Theme.of(context).colorScheme.primaryContainer,
+              Theme.of(context).colorScheme.secondaryContainer,
+              Theme.of(context).colorScheme.tertiaryContainer
+            ],
           ),
         ),
-        child: SafeArea(
-          child: SingleChildScrollView(
-            child: Padding(
-              padding: const EdgeInsets.all(30.0),
-              child: Form(
-                key: _formKey,
-                child: Column(
-                  children: [
-                    GestureDetector(
-                      onTap: _getImage,
-                      child: CircleAvatar(
-                        radius: 75,
-                        backgroundColor: Colors.white,
-                        backgroundImage:
-                            _image != null ? FileImage(_image!) : null,
-                        child: _image == null
-                            ? Icon(Icons.add_a_photo,
-                                size: 50, color: Colors.grey)
-                            : null,
-                      ),
+        child: SizedBox.expand(
+          child: Padding(
+            padding: UiConsts.PaddingAll_large,
+            child: Form(
+              key: _formKey,
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.center,
+                crossAxisAlignment: CrossAxisAlignment.center,
+                children: [
+                  GestureDetector(
+                    onTap: _getImage,
+                    child: CircleAvatar(
+                      radius: 75,
+                      backgroundColor: Colors.white,
+                      backgroundImage:
+                          _image != null ? FileImage(_image!) : null,
+                      child: _image == null
+                          ? Icon(Icons.add_a_photo,
+                              size: 50, color: Colors.grey)
+                          : null,
                     ),
-                    SizedBox(height: 30),
-                    TextFormField(
+                  ),
+                  SizedBox(height: 30),
+                  TextFormField(
                       controller: _displayNameController,
+                      autovalidateMode: AutovalidateMode.onUserInteraction,
                       decoration: InputDecoration(
-                        labelText: 'Display Name',
-                        filled: true,
-                        fillColor: Colors.white.withOpacity(0.8),
-                        border: OutlineInputBorder(
-                          borderRadius: BorderRadius.circular(30),
-                          borderSide: BorderSide.none,
-                        ),
+                          hintText: AppLocalizations.of(context)!.displayName,
+                          labelText: AppLocalizations.of(context)!.displayName,
+                          floatingLabelBehavior: FloatingLabelBehavior.auto,
+                          floatingLabelAlignment: FloatingLabelAlignment.center,
+                          floatingLabelStyle: TextStyle(
+                              color: Theme.of(context).colorScheme.onSurface),
+                          errorStyle: TextStyle(
+                              color:
+                                  Theme.of(context).colorScheme.inverseSurface),
+                          errorBorder: UnderlineInputBorder(
+                              borderRadius:
+                                  UiConsts.BorderRadiusCircular_standard,
+                              borderSide: BorderSide(
+                                color: Theme.of(context).colorScheme.error,
+                              )),
+                          focusedErrorBorder: UnderlineInputBorder(
+                            borderRadius:
+                                UiConsts.BorderRadiusCircular_standard,
+                            borderSide: BorderSide(
+                                color: Theme.of(context).colorScheme.error),
+                          ),
+                          focusedBorder: OutlineInputBorder(
+                            borderRadius:
+                                UiConsts.BorderRadiusCircular_standard,
+                            borderSide:
+                                BorderSide(color: Theme.of(context).focusColor),
+                          ),
+                          enabledBorder: OutlineInputBorder(
+                            borderRadius:
+                                UiConsts.BorderRadiusCircular_standard,
+                            borderSide: BorderSide(
+                                color: Theme.of(context).colorScheme.onPrimary),
+                          ),
+                          border: InputBorder.none),
+                      validator: FormBuilderValidators.maxLength(24)),
+                  SizedBox(height: 30),
+                  ElevatedButton(
+                    onPressed: () async => _updateProfile(),
+                    style: ElevatedButton.styleFrom(
+                      foregroundColor:
+                          Theme.of(context).colorScheme.onPrimaryContainer,
+                      backgroundColor:
+                          Theme.of(context).colorScheme.primaryContainer,
+                      padding:
+                          EdgeInsets.symmetric(horizontal: 50, vertical: 15),
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(30),
                       ),
-                      validator: (value) {
-                        if (value == null || value.isEmpty) {
-                          return 'Please enter a display name';
-                        }
-                        return null;
-                      },
                     ),
-                    SizedBox(height: 30),
-                    ElevatedButton(
-                      onPressed: _updateProfile,
-                      style: ElevatedButton.styleFrom(
-                        foregroundColor:
-                            Theme.of(context).colorScheme.onPrimaryContainer,
-                        backgroundColor:
-                            Theme.of(context).colorScheme.primaryContainer,
-                        padding:
-                            EdgeInsets.symmetric(horizontal: 50, vertical: 15),
-                        shape: RoundedRectangleBorder(
-                          borderRadius: BorderRadius.circular(30),
-                        ),
-                      ),
-                      child: Text(AppLocalizations.of(context)!
-                          .updateProfile
-                          .toUpperCase()),
-                    ),
-                  ],
-                ),
+                    child: Text(AppLocalizations.of(context)!
+                        .updateProfile
+                        .toUpperCase()),
+                  ),
+                ],
               ),
             ),
           ),
@@ -123,22 +155,13 @@ class _UpdateProfilePageState extends State<UpdateProfilePage> {
     );
   }
 
-  void _updateProfile() {
+  Future<void> _updateProfile() async {
+    context.loaderOverlay.show();
     if (_formKey.currentState!.validate()) {
-      // Implement Firebase profile update logic here
-      // Update display name
-      FirebaseAuth.instance.currentUser
-          ?.updateDisplayName(_displayNameController.text);
-
-      // Update profile picture (you'll need to implement this part)
-      // This typically involves uploading the image to Firebase Storage
-      // and then updating the user's photoURL
-
-      // Show success message and navigate back
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(content: Text('Profile updated successfully')),
-      );
-      Navigator.of(context).pop();
+      await _authWrapper.handleUpdateUserAccountProfile(
+          context, _displayNameController.text);
+      context.loaderOverlay.hide();
     }
+    context.loaderOverlay.hide();
   }
 }